Specification for tokens and create/mint/burn
This specification goes into details of what a token should contain
as information, update the ledger information regarding it (and
returns token informations instead of just names), and add a new
attribute for creating / minting / burning tokens.

It also includes account roles so a token owner can be an account
and contains roles for others.