Clarify neighborhood in status
We need clarification of neighborhood from servers, otherwise any
frontend or client might be talking to a server that does not
belong to the neighborhood.

The next step would be to have some certificate of ownership of
neighborhood (e.g. an X509), but for now this should be enough
to know that the Manifest Ledger is not the same neighborhood
as the END Labs Ledger, for example.